Conflicts:
	forms/simulation/simulationview.cpp

+++ forms/heavenlybody/heavenlybodyoverview.cpp
@@ -1,13 +1,25 @@
 #include "heavenlybodyoverview.h"
 #include "ui_heavenlybodyoverview.h"
 
-HeavenlyBodyOverview::HeavenlyBodyOverview(QWidget *parent) :
+HeavenlyBodyOverview::HeavenlyBodyOverview(QWidget *parent, HeavenlyBodyModel *heavenlyBodyModel) :
     QDialog(parent),
     ui(new Ui::HeavenlyBodyOverview)
 {
     ui->setupUi(this);
 
-    heavenlyBodyModel = new HeavenlyBodyModel();
+    QObject::connect(ui->heavenlyBodyTableView,
+                     SIGNAL(doubleClicked(QModelIndex)),
+                     this,
+                     SLOT(doubleClicked(QModelIndex)),
+                     Qt::DirectConnection);
+
+    QObject::connect(ui->heavenlyBodyTableView->selectionModel(),
+                     SIGNAL(selectionChanged(QItemSelection,QItemSelection)),
+                     this,
+                     SLOT(selectionChanged(QItemSelection,QItemSelection)),
+                     Qt::DirectConnection);
+
+    this->heavenlyBodyModel = heavenlyBodyModel;
 
     ui->heavenlyBodyTableView->setModel(heavenlyBodyModel->getHeavenlyBodyTableModel());
     heavenlyBodyModel->setSelectionModel(ui->heavenlyBodyTableView->selectionModel());
@@ -22,6 +34,21 @@ HeavenlyBodyOverview::~HeavenlyBodyOverview()
     delete ui;
 }
 
+void HeavenlyBodyOverview::doubleClicked(QModelIndex modelIndex)
+{
+    if (heavenlyBodyModel->isEntitySelected())
+    {
+        HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, true);
+        heavenlyBodyDetails->show();
+    }
+}
+
+void HeavenlyBodyOverview::selectionChanged(const QItemSelection &selected, const QItemSelection &deselected)
+{
+    ui->edit->setEnabled(selected.size() == 1);
+    ui->deleteEntity->setEnabled(selected.size() == 1);
+}
+
 void HeavenlyBodyOverview::on_add_clicked()
 {
     HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, false);
@@ -30,21 +57,32 @@ void HeavenlyBodyOverview::on_add_clicked()
 
 void HeavenlyBodyOverview::on_edit_clicked()
 {
-    HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, true);
-    heavenlyBodyDetails->show();
+    if (heavenlyBodyModel->isEntitySelected())
+    {
+        HeavenlyBodyDetails *heavenlyBodyDetails = new HeavenlyBodyDetails(this, heavenlyBodyModel, true);
+        heavenlyBodyDetails->show();
+    }
 }
 
 void HeavenlyBodyOverview::on_deleteEntity_clicked()
 {
     try
     {
-        heavenlyBodyModel->deleteEntity();
+        int result = QMessageBox::question(this,
+                                           "Delete a Heavenly Body",
+                                           QString("Would you like to delete the Heavenly Body '%1'?").arg(heavenlyBodyModel->getSelectedEntity()->getName()),
+                                           QMessageBox::Yes | QMessageBox::No);
+
+        if (result == QMessageBox::Yes)
+        {
+            heavenlyBodyModel->deleteEntity();
+        }
     }
     catch (const DeleteEntityFailedException &exception)
     {
-        QMessageBox::information(this,
-                                 "Error while deleting",
-                                 exception.getMessage() + "\n\nSQL-Error:\n" + exception.getSqlError(),
-                                 QMessageBox::Ok);
+        QMessageBox::critical(this,
+                              "Error while deleting",
+                              exception.getMessage() + "\n\nSQL-Error:\n" + exception.getSqlError(),
+                              QMessageBox::Ok);
     }
 }
